+++ test/htmlwg/index.js
@@ -1,2 +1,0 @@
-var harness = require('./harness');
-module.exports = harness('submission');