+++ test/index.js
@@ -1,3 +1,2 @@
-exports.htmlwg = require('./htmlwg');
 exports.w3c = require('./w3c');